+++ python/ray/rllib/agents/ppo/appo_policy.py
@@ -10,22 +10,18 @@ import numpy as np
 import logging
 import gym
 
-import ray
 from ray.rllib.agents.impala import vtrace
+from ray.rllib.agents.impala.vtrace_policy import _make_time_major, \
+        BEHAVIOUR_LOGITS, VTraceTFPolicy
 from ray.rllib.evaluation.postprocessing import Postprocessing
 from ray.rllib.models.action_dist import Categorical
 from ray.rllib.policy.sample_batch import SampleBatch
-from ray.rllib.policy.tf_policy_template import build_tf_policy
-from ray.rllib.policy.tf_policy import LearningRateSchedule
-from ray.rllib.utils.explained_variance import explained_variance
 from ray.rllib.evaluation.postprocessing import compute_advantages
 from ray.rllib.utils import try_import_tf
 
 tf = try_import_tf()
 
 logger = logging.getLogger(__name__)
-
-BEHAVIOUR_LOGITS = "behaviour_logits"
 
 
 class PPOSurrogateLoss(object):
@@ -163,41 +159,6 @@ class VTraceSurrogateLoss(object):
                            self.entropy * entropy_coeff)
 
 
-def _make_time_major(policy, tensor, drop_last=False):
-    """Swaps batch and trajectory axis.
-
-    Arguments:
-        policy: Policy reference
-        tensor: A tensor or list of tensors to reshape.
-        drop_last: A bool indicating whether to drop the last
-        trajectory item.
-
-    Returns:
-        res: A tensor with swapped axes or a list of tensors with
-        swapped axes.
-    """
-    if isinstance(tensor, list):
-        return [_make_time_major(policy, t, drop_last) for t in tensor]
-
-    if policy.state_in:
-        B = tf.shape(policy.seq_lens)[0]
-        T = tf.shape(tensor)[0] // B
-    else:
-        # Important: chop the tensor into batches at known episode cut
-        # boundaries. TODO(ekl) this is kind of a hack
-        T = policy.config["sample_batch_size"]
-        B = tf.shape(tensor)[0] // T
-    rs = tf.reshape(tensor, tf.concat([[B, T], tf.shape(tensor)[1:]], axis=0))
-
-    # swap B and T axes
-    res = tf.transpose(
-        rs, [1, 0] + list(range(2, 1 + int(tf.shape(tensor).shape[0]))))
-
-    if drop_last:
-        return res[:-1]
-    return res
-
-
 def build_appo_surrogate_loss(policy, batch_tensors):
     if isinstance(policy.action_space, gym.spaces.Discrete):
         is_multidiscrete = False
@@ -283,28 +244,6 @@ def build_appo_surrogate_loss(policy, batch_tensors):
     return policy.loss.total_loss
 
 
-def stats(policy, batch_tensors):
-    values_batched = _make_time_major(
-        policy, policy.value_function, drop_last=policy.config["vtrace"])
-
-    return {
-        "cur_lr": tf.cast(policy.cur_lr, tf.float64),
-        "policy_loss": policy.loss.pi_loss,
-        "entropy": policy.loss.entropy,
-        "var_gnorm": tf.global_norm(policy.var_list),
-        "vf_loss": policy.loss.vf_loss,
-        "vf_explained_var": explained_variance(
-            tf.reshape(policy.loss.value_targets, [-1]),
-            tf.reshape(values_batched, [-1])),
-    }
-
-
-def grad_stats(policy, grads):
-    return {
-        "grad_gnorm": tf.global_norm(grads),
-    }
-
-
 def postprocess_trajectory(policy,
                            sample_batch,
                            other_agent_batches=None,
@@ -337,61 +276,8 @@ def add_values_and_logits(policy):
     return out
 
 
-def validate_config(policy, obs_space, action_space, config):
-    assert config["batch_mode"] == "truncate_episodes", \
-        "Must use `truncate_episodes` batch mode with V-trace."
-
-
-def choose_optimizer(policy, config):
-    if policy.config["opt_type"] == "adam":
-        return tf.train.AdamOptimizer(policy.cur_lr)
-    else:
-        return tf.train.RMSPropOptimizer(policy.cur_lr, config["decay"],
-                                         config["momentum"], config["epsilon"])
-
-
-def clip_gradients(policy, optimizer, loss):
-    grads = tf.gradients(loss, policy.var_list)
-    policy.grads, _ = tf.clip_by_global_norm(grads, policy.config["grad_clip"])
-    clipped_grads = list(zip(policy.grads, policy.var_list))
-    return clipped_grads
-
-
-class ValueNetworkMixin(object):
-    def __init__(self):
-        self.value_function = self.model.value_function()
-        self.var_list = tf.get_collection(tf.GraphKeys.TRAINABLE_VARIABLES,
-                                          tf.get_variable_scope().name)
-
-    def value(self, ob, *args):
-        feed_dict = {
-            self.get_placeholder(SampleBatch.CUR_OBS): [ob],
-            self.seq_lens: [1]
-        }
-        assert len(args) == len(self.state_in), \
-            (args, self.state_in)
-        for k, v in zip(self.state_in, args):
-            feed_dict[k] = v
-        vf = self.get_session().run(self.value_function, feed_dict)
-        return vf[0]
-
-
-def setup_mixins(policy, obs_space, action_space, config):
-    LearningRateSchedule.__init__(policy, config["lr"], config["lr_schedule"])
-    ValueNetworkMixin.__init__(policy)
-
-
-AsyncPPOTFPolicy = build_tf_policy(
-    name="AsyncPPOTFPolicy",
-    get_default_config=lambda: ray.rllib.agents.impala.impala.DEFAULT_CONFIG,
+AsyncPPOTFPolicy = VTraceTFPolicy.with_updates(
+    name="VTraceTFPolicy",
     loss_fn=build_appo_surrogate_loss,
-    stats_fn=stats,
-    grad_stats_fn=grad_stats,
     postprocess_fn=postprocess_trajectory,
-    optimizer_fn=choose_optimizer,
-    gradients_fn=clip_gradients,
-    extra_action_fetches_fn=add_values_and_logits,
-    before_init=validate_config,
-    before_loss_init=setup_mixins,
-    mixins=[LearningRateSchedule, ValueNetworkMixin],
-    get_batch_divisibility_req=lambda p: p.config["sample_batch_size"])
+    extra_action_fetches_fn=add_values_and_logits)
